Library API

This page describes the public types and functions exported by Sunny. This documentation can be also be accessed using the Julia help system (enter ? at the Julia command prompt).

Sunny.Units

Sunny.meV_per_K

Sunny.𝒪

Sunny.𝒮

Sunny.BinnedArray

Sunny.Bond

Sunny.Crystal

Sunny.FormFactor

Sunny.ImplicitMidpoint

Sunny.IsingSampler

Sunny.LangevinHeunP

Sunny.LangevinSampler

Sunny.MetropolisSampler

Sunny.SpinInfo

Sunny.StructureFactor

Sunny.System

Sunny.WangLandau

Sunny.add_trajectory!

Sunny.all_symmetry_related_anisotropies

Sunny.all_symmetry_related_bonds

Sunny.all_symmetry_related_bonds_for_atom

Sunny.all_symmetry_related_couplings

Sunny.all_symmetry_related_couplings_for_atom

Sunny.anneal!

Sunny.browser

Sunny.calculate_structure_factor

Sunny.cell_type

Sunny.cell_volume

Sunny.connected_path

Sunny.coordination_number

Sunny.displacement

Sunny.distance

Sunny.dmvec

Sunny.enable_dipole_dipole!

Sunny.energy

Sunny.extend_periodically

Sunny.filter_visited

Sunny.forces

Sunny.get_intensities

Sunny.get_static_intensities

Sunny.get_temp

Sunny.init_bounded!

Sunny.intensity_grid

Sunny.lattice_params

Sunny.lattice_vectors

Sunny.nbasis

Sunny.print_anisotropy_as_classical_spins

Sunny.print_anisotropy_as_stevens

Sunny.print_bond

Sunny.print_site

Sunny.print_suggested_frame

Sunny.print_symmetry_table

Sunny.reference_bonds

Sunny.reset!

Sunny.run!

Sunny.sample!

Sunny.set_anisotropy!

Sunny.set_exchange!

Sunny.set_exchange_with_biquadratic!

Sunny.set_external_field!

Sunny.set_local_anisotropy!

Sunny.set_local_external_field!

Sunny.set_temp!

Sunny.step!

Sunny.subcrystal

Sunny.thermalize!

Sunny.view_crystal

Sunny.ωvals

Sunny.Units — Constant

Units.meV
Units.theory

The unit system is implicitly determined by the definition of two physical constants: the vacuum permeability $μ₀$ and the Bohr magneton $μ_B$. Temperatures are effectively measured in units of energy ($k_B = 1$) and time is effectively measured in units of inverse energy ($ħ = 1$). The default unit system, Units.meV, employs (meV, Å, tesla). Select alternatively Units.theory for a units system defined so that $μ₀ = μ_B = 1$.

See also meV_per_K

Sunny.meV_per_K — Constant

meV_per_K = 0.086173332621451774

A physical constant. Useful for converting kelvin into the default energy units, meV.

Sunny.𝒪 — Constant

𝒪[k,q]

Abstract symbols for the Stevens operators. Linear combinations of these can be used to specify the single-ion anisotropy.

Sunny.𝒮 — Constant

𝒮[1], 𝒮[2], 𝒮[3]

Abstract symbols for the spin operators. Polynomials of these can be used to specify the single-ion anisotropy.

Sunny.BinnedArray — Type

mutable struct BinnedArray{K, V}

Adaptive array that bins data. Can be used as a histogram. Just 1D now, but could use raveled index if necessary.

Sunny.Bond — Type

Bond(i, j, n)

Represents a bond between atom indices i and j. n is a vector of three integers specifying unit cell displacement in terms of lattice vectors.

Sunny.Crystal — Type

An object describing a crystallographic unit cell and its space group symmetry. Constructors are as follows:

Crystal(filename; symprec=1e-5)

Reads the crystal from a .cif file located at the path filename. The optional parameter symprec controls the precision tolerance for spacegroup symmetries.

Crystal(lat_vecs, positions; types=nothing, symprec=1e-5)

Constructs a crystal from the complete list of atom positions positions, with coordinates (between 0 and 1) in units of lattice vectors lat_vecs. Spacegroup symmetry information is automatically inferred. The optional parameter types is a list of strings, one for each atom, and can be used to break symmetry-equivalence between atoms.

Crystal(lat_vecs, positions, spacegroup_number; types=nothing, setting=nothing, symprec=1e-5)

Builds a crystal by applying symmetry operators for a given international spacegroup number. For certain spacegroups, there are multiple possible unit cell settings; in this case, a warning message will be printed, and a list of crystals will be returned, one for every possible setting. Alternatively, the optional setting string will disambiguate between unit cell conventions.

Examples

# Read a Crystal from a .cif file
Crystal("filename.cif")

# Build an FCC crystal using the primitive unit cell. The spacegroup number
# 225 is inferred.
lat_vecs = [1 1 0;
            1 0 1;
            0 1 1] / 2
positions = [[0, 0, 0]]
Crystal(lat_vecs, positions)

# Build a CsCl crystal (two cubic sublattices). By providing distinct type
# strings, the spacegroup number 221 is inferred.
lat_vecs = lattice_vectors(1, 1, 1, 90, 90, 90)
positions = [[0,0,0], [0.5,0.5,0.5]]
types = ["Na", "Cl"]
cryst = Crystal(lat_vecs, positions; types)

# Build a diamond cubic crystal from its spacegroup number 227. This
# spacegroup has two possible settings ("1" or "2"), which determine an
# overall unit cell translation.
lat_vecs = lattice_vectors(1, 1, 1, 90, 90, 90)
positions = [[1, 1, 1] / 4]
cryst = Crystal(lat_vecs, positions, 227; setting="1")

See also lattice_vectors.

Sunny.FormFactor — Method

FormFactor(atom::Int64, elem::String; g_lande=nothing)

Basic type for specifying form factor parameters. Must be provided a site within the unit cell (atom) and a string specifying the element name. This used when calling get_intensities, which requires a list of FormFactorss.

A list of supported element names is available at:

https://www.ill.eu/sites/ccsl/ffacts/ffachtml.html

The Landé g-factor may also be specified.

In more detail, the data stored in a FormFactor will be used to compute the form factor for each momentum space magnitude |k|, measured in inverse angstroms. The result is dependent on the magnetic ion species. By default, a first order form factor $f$ is returned. If the keyword g_lande is given a numerical value, then a second order form factor $F$ is returned.

It is traditional to define the form factors using a sum of Gaussian broadening functions in the scalar variable $s = |k|/4π$, where $|k|$ can be interpreted as the magnitude of momentum transfer.

The Neutron Data Booklet, 2nd ed., Sec. 2.5 Magnetic Form Factors, defines the approximation

$\langle j_l(s) \rangle = A e^{-as^2} + B e^{-bs^2} + Ce^{-cs^2} + D,$

where coefficients $A, B, C, D, a, b, c$ are obtained from semi-empirical fits, depending on the orbital angular momentum index $l = 0, 2$. For transition metals, the form-factors are calculated using the Hartree-Fock method. For rare-earth metals and ions, Dirac-Fock form is used for the calculations.

A first approximation to the magnetic form factor is

$f(s) = \langle j_0(s) \rangle$

A second order correction is given by

$F(s) = \frac{2-g}{g} \langle j_2(s) \rangle s^2 + f(s)$, where $g$ is the Landé g-factor.

Digital tables are available at:

https://www.ill.eu/sites/ccsl/ffacts/ffachtml.html

Additional references are:

Marshall W and Lovesey S W, Theory of thermal neutron scattering Chapter 6 Oxford University Press (1971)

Clementi E and Roetti C, Atomic Data and Nuclear Data Tables, 14 pp 177-478 (1974)

Freeman A J and Descleaux J P, J. Magn. Mag. Mater., 12 pp 11-21 (1979)

Descleaux J P and Freeman A J, J. Magn. Mag. Mater., 8 pp 119-129 (1978)

Sunny.ImplicitMidpoint — Type

ImplicitMidpoint(Δt::Float64; atol=1e-12) where N

Energy-conserving integrator for simulating dynamics without damping or noise. Use with the step! function to evolve a System forward by a time step of Δt:

step!(sys::System, integrator::ImplicitMidpoint)

The above function will use the spherical midpoint integration scheme for dipole systems and the Schrodinger midpoint integration scheme for SU(N) spin systems. Both integration schemes are symplectic (energy-conserving) and are appropriate for simulating dissipationless dynamics over long periods of time.

Sunny.IsingSampler — Type

IsingSampler(sys::System, kT::Float64, nsweeps::Int)

A sampler which performs the standard Metropolis Monte Carlo algorithm to sample a System at the requested temperature.

This version differs from MetropolisSampler in that each single-spin update only attempts to completely flip the spin. One call to sample! will attempt to flip each spin nsweeps times.

Before constructing, be sure that your System is initialized so that each spin points along its "Ising-like" axis.

Sunny.LangevinHeunP — Type

LangevinHeunP(kT::Float64, λ::Float64, Δt::Float64)

Projected Heun integration scheme with noise and damping. Use with the step! function to evolve a System forward by a time step of Δt:

step!(sys::System, integrator::LangevinHeunP)

If kT > 0, this will simulate dynamics in the presence of a thermal bath. λ is an empirical parameter that determines the strength of coupling to the thermal bath and sets a time scale for decorrelation, 1/λ.

Sunny.LangevinSampler — Type

LangevinSampler(integrator::LangevinHeunP, nsteps::Int)

Creates a sampler from a Langevin integrator. nsteps determines how many times step! is called using the integrator. nsteps should be selected large enough to ensure that the state of the System after integration is decorrelated with respect to its initial state.

Sunny.MetropolisSampler — Type

MetropolisSampler(sys::System, kT::Float64, nsweeps::Int)

A sampler which performs the standard Metropolis Monte Carlo algorithm to sample a System at the requested temperature.

Each single-spin update attempts to completely randomize the spin. One call to sample! will attempt to flip each spin nsweeps times.

Sunny.SpinInfo — Type

SpinInfo(atom::Int; S, g=2)

Characterizes the spin at a given atom index within the crystal unit cell. S is an integer multiple of 1/2 and gives the spin angular momentum in units of ħ. g is the g-factor or tensor, such that an angular momentum dipole $s$ produces a magnetic moment $g s$ in units of the Bohr magneton.

Sunny.StructureFactor — Method

StructureFactor(sys::System; Δt = 0.1, nω = 1, ωmax = nothing,
                    apply_g = true, ops = nothing, matrix_elems = nothing)

StructureFactor is the basic type for calculating $𝒮^{αβ}(q,ω)$, storing the results, and retrieving intensity information.

Most users should use calculate_structure_factor to generate a structure factor. However, one may also directly create a structure factor by providing a System to StructureFactor. Then, to manually generate a sample, the user must ensure that the spins of their system are drawn from an appropriate equilibrium distribution and call add_trajectory! function. This process may be repeated as necessary to generate a sufficient statistics.

The keywords specify all the parameters used in subsequent calculations and must be determined at the time that the StructureFactor is initiated. Specifically:

Δt: Sets the step size of the integrator used to calculate trajectories.

ωmax: Determines the maximum resolved energy.

nω: Determines how many energy bins to resolve between 0 and ωmax. If set to 1, Sunny will calculate a static structure factor.

apply_g: Determines whether to apply the g-factor when calculating trajectories.

ops: Enables an advanced feature for SU(N) mode, allowing the user to specify custom observables other than the three components of the dipole. To use this features, ops must be given an N×N×numops array, where the final index is used to retrieve each N×N operator.

matrix_elems: Allows the user to specify which correlation functions are calculated. This is specified with a vector of tuples. By default Sunny records all auto- and cross-correlations generated by the x, y, and z dipolar components. To retain only the xx and xy correlations, one would set matrix_elems=[(1,1), (1,2)].

process_trajectory: Specify any preprocessing that should be applied to a recorded trajectory before calculating correlations. By default, no processing is applied. Options are :symmetrize, which symmetrizes the trajectory, and :subtract_mean, which subtracts the mean value. Both options are useful for eliminating Fourier artifacts that frequently appear when simulating gapless, or nearly gapless, Hamiltonians.

If you wish to calculate a dynamical structure factor, the keyword nω must set to an integer greater than 1.

Sunny.System — Method

System(crystal::Crystal, latsize, infos, mode; units=Units.meV, seed::Int)

Construct a System of spins for a given crystal symmetry. The latsize parameter determines the number of unit cells in each lattice vector direction. The infos parameter is a list of SpinInfo objects, which determine the magnitude $S$ and $g$-tensor of each spin.

The three possible options for mode are :dipole, :SUN, and :projected. The choice :dipole restricts the description of a spin to its angular momentum dipole. The choice :SUN will expand the description of a spin to a full SU(N) coherent state, where $N = 2S + 1$. This approach captures more quantum mechanical degrees of freedom, e.g., the dynamics of quadrupolar fluctuations $⟨ŜᵅŜᵝ+ŜᵝŜᵅ⟩$. Finally, the choice :projected produces the SU(N) dynamics but projected to the space of dipoles. In practice, the distinction between :projected and :dipoles is that the former will automatically apply appropriate renormalizations to the single-ion anisotropy and biquadratic exchange interactions for maximum accuracy.

The default units system of (meV, Å, tesla) can be overridden by with the units parameter; see Units.

All spins are initially polarized in the $z$-direction.

Sunny.WangLandau — Type

mutable struct WangLandau{F<:Function}

Wang-Landau sampler. All parameters have default values that can be overwritten, but a System must be passed during construction.

Sunny.add_trajectory! — Method

add_trajectory!(sf::StructureFactor, sys::System)

When given a StructureFactor and a System, add_trajectory will use the spin state contained in the System as the initial condition for a dynamical trajectory. The correlations generated by the trajectory are calculated and accumulated into the StructureFactor. All trajectory parameters are specified in the StructureFactor.

Note that this function will not change the spin state in sys.

Sunny.all_symmetry_related_anisotropies — Method

all_symmetry_related_anisotropies(cryst, i_ref, Λ_ref)

Return two lists. The first list contains all atoms i that are symmetry equivalent to i_ref. The second list contains the appropriately transformed anisotropy matrices Λ for each site i.

Sunny.all_symmetry_related_bonds — Method

all_symmetry_related_bonds(cryst::Crystal, b::Bond)

Returns a list of all bonds that are symmetry-equivalent to bond b or its reverse.

Sunny.all_symmetry_related_bonds_for_atom — Method

all_symmetry_related_bonds_for_atom(cryst::Crystal, i::Int, b::Bond)

Returns a list of all bonds that start at atom i, and that are symmetry equivalent to bond b or its reverse.

Sunny.all_symmetry_related_couplings — Method

all_symmetry_related_couplings(cryst::Crystal, b::Bond, J)

Given a reference bond b and coupling matrix J on that bond, return a list of symmetry-equivalent bonds and a corresponding list of symmetry-transformed coupling matrices.

Sunny.all_symmetry_related_couplings_for_atom — Method

all_symmetry_related_couplings_for_atom(cryst::Crystal, i::Int, b::Bond, J)

Given a reference bond b and coupling matrix J on that bond, return a list of symmetry-equivalent bonds (constrained to start from atom i), and a corresponding list of symmetry-transformed coupling matrices.

Sunny.anneal! — Method

anneal!(sampler, temp_schedule, step_schedule)

sample! a sampler at a series of temperatures, staying at each temperature for the number of steps in step_schedule.

anneal!(sampler, temp_function::Function, num_samples)

sample! a sampler num_samples times, with the sample at timestep n drawn at a temperature temp_function(n).

Sunny.browser — Method

browser(html_str; dir)

Launch a system browser to display the provided HTML string or SunnyViewer. If a directory dir is provided, an HTML file will be written at that location.

Sunny.calculate_structure_factor — Method

calculate_structure_factor(sys::System, sampler::LangevinSampler; nsamples=10, kwargs...)

Calculates a StructureFactor containg $𝒮(q,ω)$ information with a single function call. Must be given a System and a LangevinSampler. This functions expects that the spin state contained in the System to already represent a good equilibrium sample. The function will use this state to generate an initial trajectory. After generating this first trajectory, it will call sampler before generating subsequent trajectories. In total, the function will calculate nsamples trajectories and accumulate them into a StructureFactor which is returned. The remaining keyword arguments are shared with StructureFactor.

By default, this function will calculate the static structure factor, independent of dynamics. Specify a keyword argument nω greater than 1 to measure $ω$-dependent dynamical quantities.

Sunny.cell_type — Method

cell_type(lat_vecs::Mat3)

Infer the CellType of a unit cell from its lattice vectors, i.e. the columns of lat_vecs. Report an error if lattice vectors are not in conventional form.

Sunny.cell_volume — Method

cell_volume(crystal::Crystal)

Volume of the crystal unit cell.

Sunny.connected_path — Method

connected_path(qs::Vector, density)

Takes a list of wave vectors, qs, and builds an expanded list of wave vectors that traces a path through the provided points. Also returned is a list of marker indices corresponding to the intput points. The density parameter controls the frequency of sampling.

Sunny.coordination_number — Method

coordination_number(cryst::Crystal, i::Int, b::Bond)

Returns the number times that atom i participates in a bond equivalent to b. In other words, the count of bonds that begin at atom i and that are symmetry-equivalent to b or its reverse.

Defined as length(all_symmetry_related_bonds_for_atom(cryst, i, b)).

Sunny.displacement — Method

displacement(cryst::Crystal, b::Bond)

The displacement vector $𝐫_j - 𝐫_i$ in global coordinates between atoms b.i and b.j, accounting for the integer offsets b.n between unit cells.

Sunny.distance — Method

distance(cryst::Crystal, b::Bond)

The global distance between atoms in bond b. Equivalent to norm(displacement(cryst, b)).

Sunny.dmvec — Method

dmvec(D)

Representation of the Dzyaloshinskii-Moriya interaction pseudo-vector D as an antisymmetric matrix,

  |  0   D₃ -D₂ |
  | -D₃  0   D₁ |
  |  D₂ -D₁  0  |

Useful in the context of set_exchange!.

Sunny.enable_dipole_dipole! — Method

enable_dipole_dipole!(sys::System)

Enables long-range dipole-dipole interactions,

\[ -(μ₀/4π) ∑_{⟨ij⟩} (3 (𝐌_j⋅𝐫̂_{ij})(𝐌_i⋅𝐫̂_{ij}) - 𝐌_i⋅𝐌_j) / |𝐫_{ij}|^3\]

where the sum is over all pairs of spins (singly counted), including periodic images, regularized using the Ewald summation convention. The magnetic moments are $𝐌_i = μ_B g 𝐒_i$ where $g$ is the g-factor or g-tensor, and $𝐒_i$ is the spin angular momentum dipole in units of ħ. The Bohr magneton $μ_B$ and vacuum permeability $μ_0$ are physical constants, with numerical values determined by the unit system.

Sunny.energy — Method

energy(sys::System)

Computes the total system energy.

Sunny.extend_periodically — Method

extend_periodically(sys::System{N}, mults::NTuple{3, Int64}) where N

Creates a new System identical to sys but with each dimension multiplied by the corresponding factor given in the tuple mults. The original spin configuration is simply repeated periodically.

Sunny.filter_visited — Method

Return only array elements that are visited.

Sunny.forces — Method

forces(Array{Vec3}, sys::System)

Returns the effective local field (force) at each site, $𝐁 = -∂E/∂𝐬$.

Sunny.get_intensities — Method

get_intensities(sf::StructureFactor, qs, mode; interpolation = nothing,
                   kT = nothing, formfactors = nothing, negative_energies = false)

The basic function for retrieving $𝒮(𝐪,ω)$ information from a StructureFactor. Maps an array of wave vectors qs to an array of structure factor intensities, including an additional energy index. The values of $ω$ associated with the energy index can be retrieved by calling ωvals. The three coordinates of each wave vector are measured in reciprocal lattice units, i.e., multiples of the reciprocal lattice vectors.

mode: Should be one of :trace, :perp, or :full. Determines an optional contraction on the indices $α$ and $β$ of $𝒮^{αβ}(q,ω)$. Setting trace yields $∑_α 𝒮^{αα}(q,ω)$. Setting perp will employ a polarization correction on the traced value. Setting full will return all elements $𝒮^{αβ}(q,ω)$ with contraction.

interpolation: Since $𝒮(q,ω)$ is calculated on a finite lattice, data is only available at discrete wave vectors. By default, Sunny will round a requested q to the nearest available wave vector. Linear interpolation can be applied by setting interpolation=:linear.

kT: If a temperature is provided, the intensities will be rescaled by a temperature- and ω-dependent classical-to-quantum factor. kT should be specified when making comparisons with spin wave calculations or experimental data.

formfactors: To apply form factor corrections, provide this keyword with a vector of FormFactors, one for each unique site in the unit cell. Sunny will symmetry propagate the results to all equivalent sites.

negative_energies: If set to true, Sunny will return the periodic extension of the energy axis. Most users will not want this.

Sunny.get_static_intensities — Method

get_static_intensities(sf::StructureFactor, qs, mode; kwargs...)

Return the static structure factor intensities at wave vectors qs. The functionality is very similar to get_intensities, except the returned array has dimensions identical to qs. The energy axis has been summed out.

Sunny.get_temp — Method

get_temp(sampler) :: Float64

Returns the temperature of the sampler, as kT.

Sunny.init_bounded! — Method

Initialize system to bounded range of states using throw-away WL sampling run see run!(...) for comments explaining code in init loop.

Sunny.intensity_grid — Method

intensity_grid(sf::StructureFactor, mode;
                   bzsize=(1,1,1), negative_energies = false, index_labels = false, kwargs...)

Returns intensities at discrete wave vectors for which there is exact information. Shares all keywords with get_intensities, and provides two additional options:

bzsize: Specifies the number of Brillouin zones to return, given as a 3-tuple of integers.

index_labels: If set to true, will return axis label information for the data, which may be upacked as: (; intensities, qpoints, ωs).

Sunny.lattice_params — Method

lattice_params(lat_vecs::Mat3)

Compute the lattice parameters $(a, b, c, α, β, γ)$ for the three lattice vectors provided as columns of lat_vecs. The inverse mapping is lattice_vectors.

Sunny.lattice_vectors — Method

lattice_vectors(a, b, c, α, β, γ)

Return the lattice vectors, as columns of the $3×3$ output matrix, that correspond to the conventional unit cell defined by the lattice constants $(a, b, c)$ and the angles $(α, β, γ)$ in degrees. The inverse mapping is lattice_params.

Sunny.nbasis — Method

nbasis(crystal::Crystal)

Number of basis positions (sublattices) in the unit cell.

Sunny.print_anisotropy_as_classical_spins — Method

function print_anisotropy_as_classical_spins(p)

Prints a quantum operator (e.g. linear combination of Stevens operators) as a polynomial of spin expectation values in the classical limit.

See also print_anisotropy_as_stevens.

Sunny.print_anisotropy_as_stevens — Method

function print_anisotropy_as_stevens(p; N)

Prints a quantum operator (e.g. a polynomial of the spin operators 𝒮) as a linear combination of Stevens operators. The parameter N specifies the dimension of the SU(N) representation, corresponding to quantum spin magnitude $S = (N-1)/2$. The special value N = 0 indicates the large-$S$ classical limit.

In the output, the symbol X denotes the spin operator magnitude squared. Quantum spin operators $𝒮$ of any finite dimension satisfy $X = |𝒮|^2 = S (S+1)$. To take the large-$S$ limit, however, we keep only leading order powers of $S$, such that $X = S^2$.

This function can be useful for understanding the conversions performed internally by set_anisotropy!.

For the inverse mapping, see print_anisotropy_as_classical_spins.

Sunny.print_bond — Method

print_bond(cryst::Crystal, bond::Bond; b_ref::Bond)

Prints symmetry information for bond bond. A symmetry-equivalent reference bond b_ref can optionally be provided to fix the meaning of the coefficients A, B, ...

Sunny.print_site — Method

print_site(cryst, i; R=I)

Print symmetry information for the site i, including allowed g-tensor and allowed anisotropy operator. An optional rotation matrix R can be provided to define the reference frame for expression of the anisotropy.

Sunny.print_suggested_frame — Method

printsuggestedframe(cryst, i; digits=4)

Print a suggested reference frame, as a rotation matrix R, that can be used as input to print_site(). This is useful to simplify the description of allowed anisotropies.

Sunny.print_symmetry_table — Method

print_symmetry_table(cryst::Crystal, max_dist)

Print symmetry information for all equivalence classes of sites and bonds, up to a maximum bond distance of max_dist. Equivalent to calling print_bond(cryst, b) for every bond b in reference_bonds(cryst, max_dist), where Bond(i, i, [0,0,0]) refers to a single site i.

Sunny.reference_bonds — Method

reference_bonds(cryst::Crystal, max_dist)

Returns a full list of bonds, one for each symmetry equivalence class, up to distance max_dist. The reference bond b for each equivalence class is selected according to a scoring system that prioritizes simplification of the elements in basis_for_symmetry_allowed_couplings(cryst, b).

Sunny.reset! — Method

Set all values to 0 and if specified, reset all visited flags to false.

Sunny.run! — Method

Run a Wang-Landau simulation.

Sunny.sample! — Method

sample!(sampler)

Samples sampler.sys to a new state, under the Boltzmann distribution as defined by sampler.sys.interactions.

Sunny.set_anisotropy! — Method

set_anisotropy!(sys::System, op, i::Int)

Set the single-ion anisotropy for the ith atom of every unit cell, as well as all symmetry-equivalent atoms. The parameter op may be a polynomial in symbolic spin operators 𝒮[α], or a linear combination of symbolic Stevens operators 𝒪[k,q].

The characters 𝒮 and 𝒪 can be copy-pasted from this help message, or typed at a Julia terminal using \scrS or \scrO followed by tab-autocomplete.

For systems with mode=:projected, the anisotropy operators interactions will automatically be renormalized to achieve maximum consistency with the more variationally accurate SU(N) mode.

Examples

```julia

An easy axis anisotropy in the z-direction

set_anisotropy!(sys, -D*𝒮[3]^3, i)

The unique quartic single-ion anisotropy for a site with cubic point group

symmetry

set_anisotropy!(sys, 𝒪[4,0] + 5𝒪[4,4], i)

An equivalent expression of this quartic anisotropy, up to a constant shift

set_anisotropy!(sys, 20*(𝒮[1]^4 + 𝒮[2]^4 + 𝒮[3]^4), i)

See also print_anisotropy_as_stevens.

Sunny.set_exchange! — Method

set_exchange!(sys::System, J, bond::Bond)

Sets a 3×3 spin-exchange matrix J along bond, yielding a pairwise interaction energy $𝐒_i⋅J 𝐒_j$. This interaction will be propagated to equivalent bonds in consistency with crystal symmetry. Any previous exchange interactions on these bonds will be overwritten. The parameter bond has the form Bond(i, j, offset), where i and j are atom indices within the unit cell, and offset is a displacement in unit cells.

Scalar J implies a pure Heisenberg exchange.

As a convenience, dmvec(D) can be used to construct the antisymmetric part of the exchange, where D is the Dzyaloshinskii-Moriya pseudo-vector. The resulting interaction will be $𝐃⋅(𝐒_i×𝐒_j)$.

Examples

using Sunny, LinearAlgebra

# An explicit exchange matrix
J1 = [2 3 0;
     -3 2 0;
      0 0 2]
set_exchange!(sys, J1, bond)

# An equivalent Heisenberg + DM exchange 
J2 = 2*I + dmvec([0,0,3])
set_exchange!(sys, J2, bond)

See also set_exchange_with_biquadratic!, dmvec.

Sunny.set_exchange_with_biquadratic! — Method

set_exchange_with_biquadratic!(sys::System, J, J′, bond::Bond)

Sets both quadratic and biquadratic exchange interactions along bond, yielding a pairwise energy $𝐒_i⋅J 𝐒_j + J′ (𝐒_i⋅𝐒_j)²$. These interactions will be propagated to equivalent bonds in consistency with crystal symmetry. Any previous exchange interactions on these bonds will be overwritten.

For systems with mode=:projected the biquadratic interactions will automatically be renormalized to achieve maximum consistency with the more variationally accurate SU(N) mode. This renormalization introduces a correction to the quadratic part of the exchange, which is why the two parts must be specified concurrently.

See also set_exchange!.

Sunny.set_external_field! — Method

set_external_field!(sys::System, B::Vec3)

Introduce a Zeeman coupling between all spins and an applied magnetic field B.

Sunny.set_local_anisotropy! — Method

set_local_anisotropy!(sys::System, op, idx)

Set a single-ion anisotropy for a single spin at idx, in violation of crystal periodicity. No symmetry analysis will be performed.

Sunny.set_local_external_field! — Method

set_local_external_field!(sys::System, B::Vec3, idx::CartesianIndex{4})

Introduce an applied field B localized to a single spin at idx.

Sunny.set_temp! — Method

set_temp!(sampler, kT)

Changes the temperature of the sampler to kT.

Sunny.step! — Method

step!(sys::System, integrator)

Advance the spin system forward one step using the parameters and integration scheme specified by integrator.

Sunny.subcrystal — Method

subcrystal(cryst, types) :: Crystal

Filters sublattices of a Crystal by atom types, keeping the space group unchanged.

subcrystal(cryst, classes) :: Crystal

Filters sublattices of Crystal by equivalence classes, keeping the space group unchanged.

Sunny.thermalize! — Method

thermalize!(sampler, num_samples)

sample! a sampler a given number of times.

Sunny.view_crystal — Method

view_crystal(crystal::Crystal, max_dist::Real)

Create and show crystal viewer in a VSCode or Jupyter notebook environment. The result can also be displayed using browser().

Sunny.ωvals — Method

ωvals(sf::StructureFactor)

Return the ω-values associated with the energy index of a StructureFactor.

<!–

Plotting

To reduce package load times, Sunny plotting functions are initially hidden, and only become available when the user explicitly executes "using GLMakie". It is a good idea to check that the GLMakie installation is working correctly (execute "] test GLMakie" from the Julia REPL).

plot_lattice
plot_spins
plot_bonds
plot_all_bonds
anim_integration
live_integration
live_langevin_integration

–>