Library

Here, we document all publically exposed types and methods in our Module. Developers may be interested is further documentation of the Internals.

Our package makes extensive usage of StaticArrays.jl. In particular, throughout the documentation we make use of aliases Vec3 = SVector{3, Float64}, Mat3 = SMatrix{3, 3, Float64, 9}. Additionally, some features and internals utilize OffsetArrays.jl for pleasant interfaces.

Geometry definition

FastDipole.Symmetry.Crystal — Type

Crystal

A type holding all geometry and symmetry information needed to represent a three-dimensional crystal.

Missing docstring.

Missing docstring for Crystal(::SMatrix{3, 3, Float64, 9}, ::Vector{SVector{3, Float64}}, ::Vector{String}; symprec). Check Documenter's build log for details.

FastDipole.Symmetry.Crystal — Method

Crystal(lattice::Lattice)

Construct a Crystal using geometry information in lattice, inferring symmetry information.

FastDipole.Symmetry.Crystal — Method

Crystal(filename::AbstractString; symprec=1e-5)

Parse a Crystal from a .cif file located at the path of filename.

FastDipole.Symmetry.subcrystal — Function

subcrystal(cryst, species) :: Crystal

Filter sublattices of a Crystal by species, keeping the symmetry group of the original Crystal.

subcrystal(cryst, equiv_idxs) :: Crystal

Filter sublattices of a Crystal by a list of indexes into cryst.equiv_atoms, keeping the symmetry group of the original Crystal.

FastDipole.Lattice — Type

Lattice{D, L, Db}

A type holding geometry information about a lattice in a simulation box. The type parameter D represents the dimensionality of the Lattice, while the others must satisfy L = D^2, Db = D + 1.

These other type parameters must be in the definition for technical reasons, but are inferred without needing them explicitly provided. For example, see the Lattice{D} constructor.

FastDipole.Lattice — Method

Lattice{D}(lat_vecs, basis_vecs, species, latsize)

Construct a D-dimensional Lattice.

Arguments

lat_vecs: A matrix where the lattice vectors form the columns. Must be convert-able into SMatrix{D, D, Float64, D^2}.

basis_vecs: A Vector of basis positions of sites within the unit cell, given in fractional coordinates. Each element must be convert-able into SVector{D, Float64}.

species::Vector{String}: A list of atomic species identifiers for each site. Equivalent sites should have the same identifier.

latsize: Specifies the number of unit cells extending along each lattice vector. Must be convert-able into SVector{D, Int}.

FastDipole.volume — Function

Compute the volume of the full simulation box.

FastDipole.eachcellindex — Function

Produce an iterator over all unit cell indices.

FastDipole.gen_reciprocal — Function

Generate a ReciprocalLattice for a given Lattice

FastDipole.lattice_vectors — Function

lattice_vectors(a, b, c, α, β, γ) :: Mat3

Compute a set of lattice vectors (forming the columns of the result), specified by a given set of lattice parameters $(a, b, c, α, β, γ)$.

FastDipole.lattice_params — Function

lattice_params(lat_vecs::Mat3)

Compute the lattice parameters $(a, b, c, α, β, γ)$ from a set of lattice vectors, which form the columns of lat_vecs.

lattice_params(lattice::Lattice{3})

Symmetry analysis

FastDipole.Symmetry.Bond — Type

Bond{D}

Represents a class of bond between pairs of sites in a D-dimensional crystal, separated from each other by a certain vector.

FastDipole.Symmetry.allowed_J — Function

allowed_J(cryst::Crystal, b::Bond{3}; digits=2, tol=1e-4)

Given a bond b, returns a Matrix{String} representing the allowed form of a bilinear exchange interaction matrix on this bond, given the symmetry constraints of cryst.

FastDipole.Symmetry.print_bond_table — Function

print_bond_table(cryst::Crystal, max_dist)

Pretty-prints a table of all symmetry classes of bonds present in cryst, up to a maximum bond length of max_dist in absolute coordinates. For each class, a single "canonical" bond is shown, with i, j being the two sublattices connected, and n being the displacement vector in units of the lattice vectors.

FastDipole.Symmetry.all_symmetry_related_bonds — Function

all_symmetry_related_bonds(cryst::Crystal, b_ref::Bond{3}) :: Vector{Bond{3}}

Construct a list of all bonds which are symmetry-equivalent to the reference bond b_ref within `cryst.

FastDipole.Symmetry.all_symmetry_related_interactions — Function

all_symmetry_related_interactions(cryst::Crystal, b_ref::Bond{3}, J_ref::Mat3) :: Tuple{Vector{Bond3}, Vector{Mat3}}

Given a reference bond b_ref and exchange matrix J_ref on that bond, construct lists of all symmetry-equivalent bonds and their respective transformed exchange matrices.

Interactions

FastDipole.ExternalField — Type

ExternalField(B::Vec3)

Defines an external field acting on each spin, specifically the term

\[ -∑_i 𝐁 ⋅ 𝐒_i\]

FastDipole.OnSite — Type

OnSite(J::Vec3, label="OnSite")

Construct an on-site anisotropy term specified by the vector J. Specifically, the term

\[ ∑_i ∑_α J_α S_{i, α}^2\]

Equivalent to DiagonalCoupling with bonds = [(0,0,0)], but faster.

FastDipole.Heisenberg — Type

Heisenberg{D}

Defines a Heisenberg interaction on a D-dimensional lattice. Specifically, the term

\[ J ∑_{⟨ij⟩} 𝐒_i ⋅ 𝐒_j \]

where $⟨ij⟩$ indicates a sum over all bonds in the lattice of a certain symmetry equivalence class, with each bond appearing only once.

FastDipole.Heisenberg — Method

Heisenberg(J::Float64, cryst::Crystal, bond::Bond{D}, label="Heisenberg")

Construct a Heisenberg{D} interaction of strength J acting on all bonds symmetry-equivalent to bond in cryst.

FastDipole.DiagonalCoupling — Type

DiagonalCoupling{D}

Defines a diagonal exchange interaction on a D-dimensional lattice. Specifically, the term

\[ ∑_{⟨ij⟩} ∑_α J_α S_{i,α} ⋅ S_{j,α} \]

where $⟨ij⟩$ indicates a sum over all bonds in the lattice of a certain symmetry equivalence class, with each bond appearing only once.

Warning: This type currently (and incorrectly) assumes that $𝐉$ takes the same form on every bond within the equivalency class.

FastDipole.DiagonalCoupling — Method

DiagonalCoupling(J::Vec3, cryst::Crystal, bond::Bond{D}, label="Heisenberg")

Construct a DiagonalCoupling{D} interaction with diagonal elements specified by J, acting on all bonds symmetry-equivalent to bond in cryst.

FastDipole.GeneralCoupling — Type

GeneralCoupling{D}

Defines a general exchange interaction on a D-dimensional lattice. Specifically, the term

\[ ∑_{⟨ij⟩} 𝐒_i^⊤ J 𝐒_j\]

where $⟨ij⟩$ indicates a sum over all bonds in the lattice of a certain symmetry equivalence class, with each bond appearing only once. $J$ is a $3 × 3$ matrix which may vary from bond to bond, under symmetry constraints.

FastDipole.GeneralCoupling — Method

GeneralCoupling(J::Mat3, cryst::Crystal, bond::Bond{D}, label="GenJ")

Construct a GeneralCoupling{D} interaction specified by the J matrix on the given bond in cryst. Automatically propagates the correctly transformed exchange matrices to all symmetry-equivalent bonds.

FastDipole.DipoleReal — Type

Dipole-dipole interactions computed in real-space. DipoleFourier should be preferred in actual simulations, but this type persists as a cross-check to test the Fourier-space calculations.

FastDipole.DipoleFourier — Type

Dipole-dipole interactions computed in Fourier-space. Should produce identical results (up to numerical precision) as DipoleReal, but is asymptotically faster.

FastDipole.Hamiltonian — Type

Hamiltonian{D}

Defines a Hamiltonian for a D-dimensional spin system.

FastDipole.Hamiltonian — Method

Hamiltonian{D}(ints)
Hamiltonian{D}(ints...)

Construct a Hamiltonian{D} from a list of Interaction's.

System definition

FastDipole.ChargeSystem — Type

Defines a collection of charges. Currently primarily used to test ewald summation calculations.

FastDipole.ChargeSystem — Method

ChargeSystem(lat::Lattice)

Construct a ChargeSystem on the given lattice, initialized to all zero charges.

Random.rand! — Method

rand!(sys::ChargeSystem)

Sets charges to random values uniformly drawn from $[-1, 1]$, then shifted to charge-neutrality.

FastDipole.SpinSystem — Type

Defines a collection of spins, as well as the Hamiltonian they interact under. This is the main type to interface with most of the package.

FastDipole.SpinSystem — Method

SpinSystem(lattice::Lattice, ℋ::Hamiltonian, S=1//1)

Construct a SpinSystem with spins of magnitude S residing on the given lattice, and interactions given by ℋ. Initialized to all spins pointing along the $+𝐳̂$ direction.

Random.rand! — Method

rand!(sys::SpinSystem)

Sets spins randomly sampled on the unit sphere.

FastDipole.energy — Function

energy(sys::SpinSystem)

Computes the energy of the system under sys.hamiltonian.

Sampling

FastDipole.LangevinSampler — Method

LangevinSampler(sys, kT, α, Δt, nsteps)

Creates a LangevinSampler which samples the spin system's Hamiltonian using Langevin dynamics at a temperature kT, damping coefficient α, and producing a new sample by integrating with nsteps timesteps of size Δt.

FastDipole.MetropolisSampler — Type

MetropolisSampler(sys::SpinSystem, kT::Float64, nsweeps::Int)

A sampler which performs the standard Metropolis Monte Carlo algorithm to sample a SpinSystem at the requested temperature.

Each single-spin update attempts to move the spin to a random position on the unit sphere. One call to sample! will attempt to flip each spin nsweeps times.

FastDipole.set_temp! — Function

set_temp!(sampler, kT)

Changes the temperature of the sampler to kT.

FastDipole.sample! — Function

sample!(sampler)

Samples sampler.system to a new state, under the Boltzmann distribution as defined by sampler.system.hamiltonian.

FastDipole.thermalize! — Function

thermalize!(sampler, num_samples)

sample! a sampler a given number of times.

FastDipole.anneal! — Function

anneal!(sampler, temp_schedule, step_schedule)

sample! a sampler at a series of temperatures, staying at each temperature for the number of steps in step_schedule.

anneal!(sampler, temp_function, num_samples)

sample! a sampler num_samples times, with the sample at timestep n drawn at a temperature temp_function(n).

Structure factor calculations

FastDipole.structure_factor — Function

structure_factor(sys, sampler; num_samples, dynΔt, meas_rate, num_meas
                                bz_size, therm_samples, verbose)

Measures the full structure factor tensor of a spin system, for the requested range of 𝐪-space. Returns $𝒮^{αβ}(𝐪, ω) = ⟨S^α(𝐪, ω) S^β(𝐪, ω)^∗⟩$, which is an array of shape [3, 3, Q1, ..., Qd, T] where Qi = max(1, bz_size_i * L_i)

num_samples sets the number of thermodynamic samples to measure and average across from sampler. dynΔt sets the integrator timestep during dynamics, and meas_rate sets how often snapshots are recorded during dynamics. The sampler is sampled therm_samples times before any measurements are made.

The maximum frequency sampled is ωmax = 2π / (dynΔt * meas_rate), and the frequency resolution is set by num_meas (the number of spin snapshots measured during dynamics).

Indexing the result at (α, β, q1, ..., qd, w) gives $S^{αβ}(𝐪, ω)$ at 𝐪 = q1 * a⃰ + q2 * b⃰ + q3 * c⃰, and ω = maxω * w / T, where a⃰, b⃰, c⃰ are the reciprocal lattice vectors of sys.lattice.

Allowed values for the qi indices lie in -div(Qi, 2):div(Qi, 2, RoundUp), and allowed values for the w index lie in 0:T-1.

FastDipole.dipole_factor — Function

dipole_factor(struct_factor, lattice)

Applies the neutron dipole factor, reducing the structure factor tensor to the observable quantities. Specifically, performs the contraction: $𝒮(𝐪, ω) = ∑_{αβ} (δ_{αβ} - 𝐪̂_α 𝐪̂_β) 𝒮^{αβ}(𝐪, ω)$.

struct_factor should be of size 3 × 3 × Q1 × ⋯ × QN × T

Returns a real array of size Q1 × ⋯ × QN × T

Plotting

All plotting functions rely on a successful user installation of GLMakie, which is dependent on the correct video drivers being installed on your system. To ensure your installation is working correctly, please press ] in a Julia REPL to access the package manager, then execute test GLMakie.

FastDipole.plot_lattice — Function

plot_lattice(lattice; kwargs...)

Plot a Lattice{2} or Lattice{3}. kwargs are passed to GLMakie.scatter!.

FastDipole.plot_spins — Function

plot_spins(sys::SpinSystem; linecolor=:grey, arrowcolor=:red, linewidth=0.1,
                            arrowsize=0.3, arrowlength=1.0, kwargs...)

Plot the spin configuration defined by sys. kwargs are passed to GLMakie.arrows.

FastDipole.plot_bonds — Function

plot_bonds(lattice, ints; bondwidth=4, kwargs...)

Plot a list of pair interactions defined on a Lattice{2} or Lattice{3}. $kwargs$ are passed to plot_lattice!.

plot_bonds(sys::SpinSystem; kwargs...)

Plot all pair interactions appearing in sys.hamiltonian, on the lattice defined by sys.lattice.

FastDipole.plot_all_bonds — Function

plot_all_bonds(lattice::Lattice{3}, max_dist; kwargs...)

Plot all bond equivalency classes present in lattice up to a maximum bond length of max_dist. kwargs are passed to plot_bonds.

FastDipole.anim_integration — Function

anim_integration(sys, fname, steps_per_frame, Δt, nframes; kwargs...)

Produce an animation of constant-energy Landau-Lifshitz dynamics of the given sys.

Arguments:

sys::SpinSystem: The spin system to integrate.

fname::String: The path to save the animation to.

steps_per_frame::Int: The number of integration steps to take per frame.

Δt::Float64: The integration timestep size.

nframes::Int: The number of frames to produce in the animation.

Other keyword arguments are passed to GLMakie.arrows.

FastDipole.live_integration — Function

live_integration(sys, steps_per_frame, Δt; kwargs...)

Performs endless live constant-energy Landau-Lifshitz integration in an interactive window.

FastDipole.live_langevin_integration — Function

live_langevin_integration(sys, steps_per_frame, Δt, kT; α=0.1, kwargs...)

Performs endless live Langevin Landau-Lifshitz integration in an interactive window.

Ewald summation

These functions are not intended to be used by typical users, who instead should utilize dipole interactions purely through DipoleReal and (preferably) DipoleFourier. However, developers may find the following documentation of the internals useful.

FastDipole.ewald_sum_monopole — Function

ewald_sum_monopole(sys::ChargeSystem; η=1.0, extent=10)

Performs ewald summation to calculate the potential energy of a system of monopoles with periodic boundary conditions.

Specifically, computes:

\[ U = \frac{1}{2} \sum_{i,j} q_i q_j \left\{ \sum_{\boldsymbol{n}}^{'} \frac{\mathrm{erfc}(η|\boldsymbol{r}_{ij} + \boldsymbol{n}|)}{|\boldsymbol{r}_{ij} + \boldsymbol{n}|} + \frac{4π}{V} \sum_{\boldsymbol{k} \ne 0} \frac{e^{-k^2/4η^2}}{k^2} e^{i\boldsymbol{k}\cdot\boldsymbol{r}_{ij}} \right\} - \frac{π}{2Vη^2}Q^2 - \frac{η}{\sqrt{π}} \sum_{i} q_i^2\]

FastDipole.ewald_sum_dipole — Function

ewald_sum_dipole(sys::SpinSystem; η=1.0, extent=10)

Performs ewald summation to calculate the potential energy of a system of dipoles with periodic boundary conditions.

Missing docstring.

Missing docstring for precompute_monopole_ewald_compressed. Check Documenter's build log for details.

Missing docstring.

Missing docstring for precompute_dipole_ewald_compressed. Check Documenter's build log for details.

Missing docstring.

Missing docstring for contract_monopole_compressed. Check Documenter's build log for details.

Missing docstring.

Missing docstring for contract_dipole_compressed. Check Documenter's build log for details.

Integrators

These functions are not intended to be used by typical users, who instead should instead perform dynamics either using LangevinSampler or implicitly in Structure factor calculations. However, advanced users and developers may want direct access to an interface to perform dynamics integrations.

FastDipole.HeunP — Type

Integrator for a 2nd-order energy-conserving Heun + projection scheme

FastDipole.LangevinHeunP — Type

Integrator implementing Langevin dynamics using a 2nd-order Heun + projection scheme

FastDipole.evolve! — Function

evolve!(integrator, Δt)

Performs a single integrator timestep of size Δt.